% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/estimateActivity.R
\name{estimateActivity}
\alias{estimateActivity}
\alias{estimateActivity,matrix,matrix-method}
\title{estimateActivity}
\usage{
estimateActivity(counts, X, ...)

\S4method{estimateActivity}{matrix,matrix}(
  counts,
  X,
  U = NULL,
  model = "poisson",
  alpha = NULL,
  alphaScale = 1,
  maxIter = 500,
  plot = FALSE,
  verbose = FALSE,
  epsilon = 0.01,
  iterOLS = 0,
  lassoFamily = "gaussian",
  repressions = TRUE,
  rho_t = NULL,
  sparse = TRUE
)
}
\arguments{
\item{counts}{Gene expression counts of dimensions G x n.}

\item{X}{Gene regulatory network of dimensions G x T.}

\item{...}{parameters including:}

\item{U}{Design matrix, of dimensions n x p.}

\item{model}{The model to use. Options are \code{"poisson"}, \code{"dirMult"},
and \code{"dirMultEmpBayes"}. Defaults to \code{"poisson"}.}

\item{alpha}{Prior belief in GRN edges.}

\item{alphaScale}{Scaling factor to tune the belief of the prior versus the
observed dataset. Setting \code{alphaScale=1} (the default) means assuming equal belief to
the prior as to the observed dataset. Setting \code{alphaScale=1/2}
(\code{alphaScale=2}) means you belief the prior half (twice) as much as the
observed data. Note that providing a very low number will result in many
estimate TF activities to equal zero. Indeed, due to the -1 in the
calculation for mu_gtc, many will be negative and will hence be reset to zero.}

\item{maxIter}{Maximum number of iterations.}

\item{plot}{Logical: should log-likelihood be plotted across iterations?}

\item{verbose}{Logical: should log-likelihood and iteration be printed?}

\item{epsilon}{Log-likelihood convergence criterion.}

\item{iterOLS}{Number of OLS iterations to use to set up sparse initialization.
Defaults to 0.}

\item{lassoFamily}{The family for the sparse initialization. Defaults to \code{"gaussian"}.}

\item{repressions}{Logical: should repressions be taken into account?}

\item{rho_t}{Thinning factor to account for repressions. \code{NULL} by default.}

\item{sparse}{Logical: should sparse initialization be used?}
}
\description{
Estimate TF activity given a gene regulatory network.
}
\examples{
x=2
}
