% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/tfCounts.R
\name{tfCounts}
\alias{tfCounts}
\alias{tfCounts,matrix,ANY,matrix-method}
\title{tfCounts}
\usage{
tfCounts(counts = "matrix", pi_gtc, mu_gtc, ...)

\S4method{tfCounts}{matrix,ANY,matrix}(counts, pi_gtc = NULL, mu_gtc, design = NULL)

\S4method{tfCounts}{matrix,array,ANY}(counts, pi_gtc, mu_gtc = NULL, design = NULL)
}
\arguments{
\item{counts}{Gene expression counts.}

\item{mu_gtc}{Output from estimateActivity.}

\item{...}{parameters including:}

\item{design}{Design matrix.}
}
\value{
A matrix of estimated counts, with rows corresponding to transcription
factors and columns to cells.
}
\description{
Get counts produced by each transcription factor.
}
\examples{
counts <- matrix(rpois(n= 1e4, lambda=4), nrow=100, ncol=100)
X <- matrix(0, nrow=100, ncol=12)
X[cbind(sample(100, size=250, replace=TRUE), sample(12, size=250, replace=TRUE))] <- 1
rownames(X) <- rownames(counts) <- paste0("gene",1:100)
act <- estimateActivity(counts, X)
Y_ti <- tfCounts(act$mu_gtc, counts)
}
